2. Making DeepFakes more spurious: evading deep face forgery detection via trace removal attack/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
appdirs=1.4.4=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pyhd3eb1b0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=19.10b0=py_0
blas=1.0=mkl
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py38h06a4308_2
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2021.10.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.0=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h2531618_2
flake8=3.9.2=pyhd3eb1b0_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2021.10.1=pyhd3eb1b0_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.6.1=py38he5a9106_0
ipython=7.30.1=py38h578d9bd_0
jedi=0.18.1=py38h578d9bd_0
jpeg=9b=h024ee3a_2
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py38h578d9bd_1
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
locket=0.2.1=py38h06a4308_1
lpips=0.1.4=pypi_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.0=py38h06a4308_0
matplotlib-base=3.5.0=py38h3ed280b_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mccabe=0.6.1=py38_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py38h06a4308_1
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=py38hd09550d_3
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.5.62=pypi_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py38h8c16a72_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.2.0=pyhd3eb1b0_0
pathspec=0.7.0=py_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
prompt-toolkit=3.0.24=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.3.1=pyhd3eb1b0_0
pygments=2.11.1=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytorch-msssim=0.2.1=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=6.0=py38h7f8727e_1
pyzmq=19.0.2=py38ha71036d_2
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
regex=2021.11.2=py38h7f8727e_0
scikit-image=0.18.3=py38h51133e4_0
scipy=1.7.1=py38h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_0
snowballstemmer=2.2.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
tifffile=2020.10.1=py38hdd07704_2
timm=0.4.12=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
torch-summary=1.4.5=pypi_0
torchattacks=3.2.4=pypi_0
torchaudio=0.7.2=py38
torchextractor=0.3.0=pypi_0
torchvision=0.8.2=py38_cu110
tornado=6.1=py38h497a2fe_1
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd8ed1ab_0
typed-ast=1.4.3=py38h7f8727e_1
typing_extensions=3.10.0.2=pyh06a4308_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
